# backend-api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "spotChat",
        "description": "Api for spotChat app in reactNative",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/places": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "List or search places",
                "description": "Returns a list of places. You can search by city, by place type, or by geographic coordinates + radius.",
                "operationId": "get_api_place_list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (pagination)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "example": 20
                        }
                    },
                    {
                        "name": "placeType",
                        "in": "query",
                        "description": "Filter by place type (example: bar, club)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "bar"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter by city name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Paris"
                        }
                    },
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Latitude for geolocation search",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 48.8566
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitude for geolocation search",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 2.3522
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Search radius in kilometers (default: 10km)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of places",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Place"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Places"
                ],
                "summary": "Create a new place",
                "description": "Creates a new place using the fields allowed in the serializer group `place:write`.",
                "operationId": "post_api_place_create",
                "requestBody": {
                    "description": "Place data to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Place"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Place created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Place"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors or invalid request data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": "This value should not be blank.",
                                                "latitude": "This value should be a valid number."
                                            }
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid data: unexpected field"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/places/{id}": {
            "get": {
                "operationId": "get_api_place_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "put": {
                "tags": [
                    "Places"
                ],
                "summary": "Update an existing place",
                "description": "Updates an existing place using the fields allowed in the serializer group `place:write`.",
                "operationId": "put_api_place_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the place to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated place data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Place"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Place updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Place"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Place not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Place not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors or invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": "This value should not be blank.",
                                                "latitude": "This value should be a valid number."
                                            }
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid data: Unexpected field."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Places"
                ],
                "summary": "Delete a place",
                "description": "Deletes an existing place by its ID.",
                "operationId": "delete_api_place_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the place to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Place deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Place deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Place not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Place not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "operationId": "patch_api_place_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/places/search": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Search places",
                "description": "Searches for places by name, description, or city. Returns up to 50 results ordered by rating.",
                "operationId": "get_api_place_search",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search text (partial match). Example: bar, club, paris...",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "cocktail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results for matching places",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Place"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing search query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Search query is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Place": {
                "type": "object"
            },
...
        }
    },
...
}